+++ hardhat.config.ts
@@ -10,6 +10,7 @@ const privateKey =
     : '123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0';
 
 const POLYGONSCAN_KEY = process.env.POLYGONSCAN_KEY;
+const ETHERSCAN_KEY = process.env.ETHERSCAN_KEY;
 
 const config: HardhatUserConfig = {
   solidity: {
@@ -40,11 +41,21 @@ const config: HardhatUserConfig = {
       gasPrice: 'auto', // Gas price settings
       accounts: [`0x${privateKey}`],
     },
+
+    ethMainnet: {
+      url: 'https://eth.llamarpc.com/', // URL of your custom network
+      chainId: 1, // The Chain ID of your custom network
+      gas: 'auto', // Gas settings
+      gasPrice: 'auto', // Gas price settings
+      accounts: [`0x${privateKey}`],
+    },
   },
 
   etherscan: {
     apiKey: {
       polygon: POLYGONSCAN_KEY,
+      eth: ETHERSCAN_KEY,
+      mainnet: ETHERSCAN_KEY,
     },
   },
 };