+++ test/ERC721Universal.ts
@@ -8,9 +8,17 @@ import { ERC721Universal } from "../typechain-types/contracts/ERC721Universal.js
 import { ERC721ReceiverMock } from "../typechain-types/contracts/tests/ERC721ReceiverMock.js";
 import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
 
+function buildTokenId(slot: string, addr: string) {
+  return ethers.toBeHex(
+    "0x" + slot + addr.substring(2),
+    32,
+  );
+}
+
 describe("ERC721Universal", function () {
   const maxBalance = 2n ** 96n;
   const defaultURI = "evochain1/collectionId/";
+  const nullAddress = ethers.toBeHex(0, 20);
 
   let addr1: HardhatEthersSigner;
   let addr2: HardhatEthersSigner;
@@ -38,7 +46,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Should report correct version of the uERC721 interface", async function () {
-    expect(await erc721.ERC721UniversalVersion()).to.equal(1);
+    expect(await erc721.ERC721UniversalVersion()).to.equal(2);
   });
 
   it("Should support the standard ERC721 interface", async function () {
@@ -49,10 +57,10 @@ describe("ERC721Universal", function () {
     await interfaceId.waitForDeployment();
 
     // Tests both via direct contract calls, as well the on-chain OpenZeppelin lib checker:
-    const specified721Id = "0x80ac58cd";
-    expect(await interfaceId.getERC721Id()).to.equal(specified721Id);
-    expect(await erc721.supportsInterface(specified721Id)).to.equal(true);
-    expect(await interfaceId.supportsInterface(await erc721.getAddress(), specified721Id)).to.equal(true);
+    const specifiedId = "0x80ac58cd";
+    expect(await interfaceId.getERC721Id()).to.equal(specifiedId);
+    expect(await erc721.supportsInterface(specifiedId)).to.equal(true);
+    expect(await interfaceId.supportsInterface(await erc721.getAddress(), specifiedId)).to.equal(true);
   });
 
   it("Should support standard ERC165 interface", async function () {
@@ -90,7 +98,6 @@ describe("ERC721Universal", function () {
 
   it("Should emit OwnershipTransferred event on deploy", async function () {
     const deployedTx = erc721.deploymentTransaction();
-    const nullAddress = ethers.toBeHex(0, 20);
     await expect(deployedTx)
       .to.emit(erc721, "OwnershipTransferred")
       .withArgs(nullAddress, addr1.address);
@@ -136,12 +143,7 @@ describe("ERC721Universal", function () {
   it("Should return correct tokenURI", async function () {
     const tokenId = 1;
     expect(await erc721.tokenURI(tokenId)).to.equal(defaultURI + 'GeneralKey(' + tokenId + ')');
-
-    const slot = "34";
-    const largeTokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const largeTokenId = buildTokenId("34", addr1.address);
     const largeTokenIdAsUint256 = ethers.toBigInt(largeTokenId);
     expect(await erc721.tokenURI(largeTokenId)).to.equal(
       defaultURI + 'GeneralKey(' + largeTokenIdAsUint256.toString() + ')',
@@ -155,46 +157,33 @@ describe("ERC721Universal", function () {
   });
 
   it("The null address is the only one that cannot own tokens", async function () {
-    const slot = "34";
-    const nullAddress = ethers.toBeHex(0, 20);
-    const tokenId = ethers.toBeHex("0x" + slot + nullAddress.substring(2), 32);
+    const tokenId = buildTokenId("34", nullAddress);
     await expect(erc721.ownerOf(tokenId))
       .to.be.revertedWithCustomError(erc721, "ERC721NonexistentToken")
       .withArgs(tokenId);
   });
 
   it("initOwner decodes as expected", async function () {
-    let slot = "111";
-    let tokenId = ethers.toBeHex("0x" + slot + addr1.address.substring(2), 32);
+    let tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.initOwner(tokenId)).to.equal(addr1.address);
 
-    slot = "";
-    tokenId = ethers.toBeHex("0x" + slot + addr1.address.substring(2), 32);
+    tokenId = buildTokenId("", addr1.address);
     expect(await erc721.initOwner(tokenId)).to.equal(addr1.address);
 
-    slot = "";
-    tokenId = ethers.toBeHex("0x" + slot + addr2.address.substring(2), 32);
+    tokenId = buildTokenId("", addr2.address);
     expect(await erc721.initOwner(tokenId)).to.equal(addr2.address);
 
-    slot = "";
-    const nullAddress = ethers.toBeHex(0, 20);
-    tokenId = ethers.toBeHex("0x" + slot + nullAddress.substring(2), 32);
+    tokenId = buildTokenId("", nullAddress);
     expect(await erc721.initOwner(tokenId)).to.equal(nullAddress);
 
     const largestSlot = ethers.toBeHex(2n ** 96n - 1n, 12);
-    tokenId = ethers.toBeHex(
-      "0x" + largestSlot.substring(2) + addr2.address.substring(2),
-      32,
-    );
+    tokenId = buildTokenId(largestSlot.substring(2), addr2.address);
+
     expect(await erc721.initOwner(tokenId)).to.equal(addr2.address);
   });
 
   it("Owner of the asset should be able to transfer asset", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     expect(await erc721.balanceOf(addr1.address)).to.equal(maxBalance);
@@ -222,12 +211,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Owner of the asset should be able to burn the asset", async function () {
-    const nullAddress = ethers.toBeHex(0, 20);
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     expect(await erc721.balanceOf(addr1.address)).to.equal(maxBalance);
@@ -248,12 +232,7 @@ describe("ERC721Universal", function () {
   });
 
   it("burn can be executed by approved operator", async function () {
-    const nullAddress = ethers.toBeHex(0, 20);
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
     expect(await erc721.balanceOf(addr1.address)).to.equal(maxBalance);
 
@@ -269,11 +248,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Asset cannot be burned by address that is not its owner", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     await expect(erc721.connect(addr2).burn(tokenId))
@@ -282,12 +257,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Asset cannot be burned twice", async function () {
-    const nullAddress = ethers.toBeHex(0, 20);
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     await expect(erc721.connect(addr1).burn(tokenId))
@@ -300,12 +270,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Burned asset cannot be transferred", async function () {
-    const nullAddress = ethers.toBeHex(0, 20);
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(erc721.connect(addr1).burn(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, nullAddress, tokenId);
@@ -324,12 +289,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Burned asset has no owner - query must fail", async function () {
-    const nullAddress = ethers.toBeHex(0, 20);
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(erc721.connect(addr1).burn(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, nullAddress, tokenId);
@@ -346,12 +306,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Burned asset has no tokenURI - query must fail", async function () {
-    const nullAddress = ethers.toBeHex(0, 20);
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(erc721.connect(addr1).burn(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, nullAddress, tokenId);
@@ -368,14 +323,9 @@ describe("ERC721Universal", function () {
   });
 
   it("Owner of the asset cannot transfer to null address via transfer method", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
-    const nullAddress = ethers.toBeHex(0, 20);
     await expect(
       erc721.connect(addr2).transferFrom(addr2.address, nullAddress, tokenId),
     )
@@ -384,11 +334,7 @@ describe("ERC721Universal", function () {
   });
 
   it("User should not be able to transfer an asset that he does not own", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     await expect(
@@ -399,11 +345,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Owner of the asset should be able to do safe transfer of asset", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     const ERC721ReceiverMock =
@@ -426,11 +368,7 @@ describe("ERC721Universal", function () {
   });
 
   it("Owner of the asset should be able to do safe transfer of asset with data", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     const ERC721ReceiverMock =
@@ -459,11 +397,7 @@ describe("ERC721Universal", function () {
   });
 
   it("When Owner of the asset does safe transfer the receiver contract reverts on call", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     const ERC721ReceiverMockFactory =
@@ -483,11 +417,7 @@ describe("ERC721Universal", function () {
   });
 
   it("When Owner of the asset does safe transfer with data the receiver contract reverts on call", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.ownerOf(tokenId)).to.equal(addr1.address);
 
     const ERC721ReceiverMockFactory =
@@ -546,10 +476,10 @@ describe("ERC721UpdatableBaseURI", function () {
     await interfaceId.waitForDeployment();
 
     // Tests both via direct contract calls, as well the on-chain OpenZeppelin lib checker:
-    const specified721Id = "0xb8382a4b";
-    expect(await interfaceId.getERC721UpdatableBaseURIId()).to.equal(specified721Id);
-    expect(await erc721.supportsInterface(specified721Id)).to.equal(true);
-    expect(await interfaceId.supportsInterface(await erc721.getAddress(), specified721Id)).to.equal(true);
+    const specifiedId = "0xb8382a4b";
+    expect(await interfaceId.getERC721UpdatableBaseURIId()).to.equal(specifiedId);
+    expect(await erc721.supportsInterface(specifiedId)).to.equal(true);
+    expect(await interfaceId.supportsInterface(await erc721.getAddress(), specifiedId)).to.equal(true);
   });
 
   it("baseURI cannot be updated by address that is not owner", async function () {
@@ -605,6 +535,7 @@ describe("ERC721UpdatableBaseURI", function () {
 
 describe("ERC721Broadcast", function () {
   const defaultURI = "evochain1/collectionId/";
+  const nullAddress = ethers.toBeHex(0, 20);
 
   let addr1: HardhatEthersSigner;
   let addr2: HardhatEthersSigner;
@@ -635,27 +566,19 @@ describe("ERC721Broadcast", function () {
     await interfaceId.waitForDeployment();
 
     // Tests both via direct contract calls, as well the on-chain OpenZeppelin lib checker:
-    const specified721Id = "0x9430f0b8";
-    expect(await interfaceId.getERC721BroadcastId()).to.equal(specified721Id);
-    expect(await erc721.supportsInterface(specified721Id)).to.equal(true);
-    expect(await interfaceId.supportsInterface(await erc721.getAddress(), specified721Id)).to.equal(true);
+    const specifiedId = "0x8f8376c4";
+    expect(await interfaceId.getERC721BroadcastId()).to.equal(specifiedId);
+    expect(await erc721.supportsInterface(specifiedId)).to.equal(true);
+    expect(await interfaceId.supportsInterface(await erc721.getAddress(), specifiedId)).to.equal(true);
   });
 
   it("wasEverTransferred returns false on non-transferred assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.wasEverTransferred(tokenId)).to.equal(false);
   });
 
   it("wasEverTransferred returns true on transferred assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(
       erc721.connect(addr1).transferFrom(addr1.address, addr2.address, tokenId),
     )
@@ -665,12 +588,7 @@ describe("ERC721Broadcast", function () {
   });
 
   it("wasEverTransferred returns true on burned assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
-    const nullAddress = ethers.toBeHex(0, 20);
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(erc721.connect(addr1).burn(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, nullAddress, tokenId);
@@ -678,11 +596,7 @@ describe("ERC721Broadcast", function () {
   });
 
   it("wasEverTransferred returns true on burned assets after having been transferred", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     expect(await erc721.wasEverTransferred(tokenId)).to.equal(false);
     await expect(
       erc721.connect(addr1).transferFrom(addr1.address, addr2.address, tokenId),
@@ -690,7 +604,6 @@ describe("ERC721Broadcast", function () {
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, addr2.address, tokenId);
     expect(await erc721.wasEverTransferred(tokenId)).to.equal(true);
-    const nullAddress = ethers.toBeHex(0, 20);
     await expect(erc721.connect(addr2).burn(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(addr2.address, nullAddress, tokenId);
@@ -698,52 +611,127 @@ describe("ERC721Broadcast", function () {
   });
 
   it("broadcastMint works on non-transferred asset, and emits expected event", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
-    const nullAddress = ethers.toBeHex(0, 20);
+    const tokenId = buildTokenId("111", addr1.address);
     // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
     await expect(erc721.connect(addr2).broadcastMint(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(nullAddress, addr1.address, tokenId);
   });
 
-  it("broadcastSelfTransfer work on non-transferred asset, and emits expected event", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+  it("broadcastMintBatch works on non-transferred assets, and emits expected events", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    const tokenId2 = buildTokenId("222", addr2.address);
+    // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
+    await expect(erc721.connect(addr2).broadcastMintBatch([tokenId, tokenId2]))
+      .to.emit(erc721, "Transfer")
+      .withArgs(nullAddress, addr1.address, tokenId)
+      .and.to.emit(erc721, "Transfer")
+      .withArgs(nullAddress, addr2.address, tokenId2);
+  });
+
+  it("broadcastSelfTransfer works on non-transferred asset, and emits expected event", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
     // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
     await expect(erc721.connect(addr2).broadcastSelfTransfer(tokenId))
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, addr1.address, tokenId);
+   });
+
+   it("broadcastSelfTransferBatch works on non-transferred assets, and emits expected events", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    const tokenId2 = buildTokenId("222", addr2.address);
+    // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
+    await expect(erc721.connect(addr2).broadcastSelfTransferBatch([tokenId, tokenId2]))
+      .to.emit(erc721, "Transfer")
+      .withArgs(addr1.address, addr1.address, tokenId)
+      .and.to.emit(erc721, "Transfer")
+      .withArgs(addr2.address, addr2.address, tokenId2);
+   });
+
+   it("Standard mint on ERC721Enumerable costs gas as expected", async function () {
+    const ERC721EnumerableFactory = await ethers.getContractFactory(
+      "ERC721EnumerableMock",
+    );
+    const erc721Enum = await ERC721EnumerableFactory.deploy(
+      "laos-kitties",
+      "LAK"
+    );
+    await erc721Enum.waitForDeployment();
+    const tokenId = buildTokenId("111", addr1.address);
+    const tx = await erc721Enum.connect(addr2).mint(addr2.address, tokenId);
+    const receipt = await tx.wait();
+    expect(receipt?.gasUsed).to.equal(140683);
+  });
+   
+  it("broadcastMint cost of gas is as expected", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
+    const tx = await erc721.connect(addr2).broadcastMint(tokenId);
+    const receipt = await tx.wait();
+    expect(receipt?.gasUsed).to.equal(28140);
+  });
+
+  it("broadcastMintBatch cost of gas is as expected", async function () {
+    const tokenIds: string[] = [];
+    const nTokens = 100;
+    for (let slot = 111; slot < 111 + nTokens; slot++) {
+      const tokenId = buildTokenId(slot.toString(), addr1.address);
+      tokenIds.push(tokenId);
+    }
+    // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
+    const tx = await erc721.connect(addr2).broadcastMintBatch(tokenIds);
+    const receipt = await tx.wait();
+    expect(receipt?.gasUsed).to.equal(718563);
+  });
+
+  it("broadcastSelfTransfer cost of gas is as expected", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
+    const tx = await erc721.connect(addr2).broadcastSelfTransfer(tokenId);
+    const receipt = await tx.wait();
+    expect(receipt?.gasUsed).to.equal(28164);
+  });
+
+  it("broadcastSelfTransferBatch cost of gas is as expected", async function () {
+    const tokenIds: string[] = [];
+    const nTokens = 100;
+    for (let slot = 111; slot < 111 + nTokens; slot++) {
+      const tokenId = buildTokenId(slot.toString(), addr1.address);
+      tokenIds.push(tokenId);
+    }
+    // note that the broadcasts are sent by any address; in this example, the address is not the owner of the asset
+    const tx = await erc721.connect(addr2).broadcastSelfTransferBatch(tokenIds);
+    const receipt = await tx.wait();
+    expect(receipt?.gasUsed).to.equal(721319);
   });
 
   it("broadcastMint reverts on transferred assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(
       erc721.connect(addr1).transferFrom(addr1.address, addr2.address, tokenId),
     )
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, addr2.address, tokenId);
     await expect(erc721.connect(addr2).broadcastMint(tokenId))
+      .to.be.revertedWithCustomError(erc721, "ERC721UniversalAlreadyTransferred")
+      .withArgs(tokenId);
+  });  
+
+  it("broadcastMintBatch reverts on at least one transferred asset", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    const tokenId2 = buildTokenId("222", addr1.address);
+    await expect(
+      erc721.connect(addr1).transferFrom(addr1.address, addr2.address, tokenId),
+    )
+      .to.emit(erc721, "Transfer")
+      .withArgs(addr1.address, addr2.address, tokenId);
+    await expect(erc721.connect(addr2).broadcastMintBatch([tokenId, tokenId2]))
       .to.be.revertedWithCustomError(erc721, "ERC721UniversalAlreadyTransferred")
       .withArgs(tokenId);
   });  
 
   it("broadcastSelfTransfer reverts on transferred assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(
       erc721.connect(addr1).transferFrom(addr1.address, addr2.address, tokenId),
     )
@@ -754,13 +742,21 @@ describe("ERC721Broadcast", function () {
       .withArgs(tokenId);
   });  
 
+  it("broadcastSelfTransferBatch reverts on at least one transferred asset", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    const tokenId2 = buildTokenId("222", addr1.address);
+    await expect(
+      erc721.connect(addr1).transferFrom(addr1.address, addr2.address, tokenId),
+    )
+      .to.emit(erc721, "Transfer")
+      .withArgs(addr1.address, addr2.address, tokenId);
+    await expect(erc721.connect(addr2).broadcastSelfTransferBatch([tokenId, tokenId2]))
+      .to.be.revertedWithCustomError(erc721, "ERC721UniversalAlreadyTransferred")
+      .withArgs(tokenId);
+  });  
+
   it("broadcastMint reverts on burned assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
-    const nullAddress = ethers.toBeHex(0, 20);
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(
       erc721.connect(addr1).burn(tokenId),
     )
@@ -771,19 +767,40 @@ describe("ERC721Broadcast", function () {
       .withArgs(tokenId);
   });
 
+  it("broadcastMintBatch reverts on at least one burned asset", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    const tokenId2 = buildTokenId("222", addr1.address);
+    await expect(
+      erc721.connect(addr1).burn(tokenId),
+    )
+      .to.emit(erc721, "Transfer")
+      .withArgs(addr1.address, nullAddress, tokenId);
+    await expect(erc721.connect(addr2).broadcastMintBatch([tokenId, tokenId2]))
+      .to.be.revertedWithCustomError(erc721, "ERC721UniversalAlreadyTransferred")
+      .withArgs(tokenId);
+  });
+
   it("broadcastSelfTransfer reverts on burned assets", async function () {
-    const slot = "111";
-    const tokenId = ethers.toBeHex(
-      "0x" + slot + addr1.address.substring(2),
-      32,
-    );
-    const nullAddress = ethers.toBeHex(0, 20);
+    const tokenId = buildTokenId("111", addr1.address);
     await expect(
       erc721.connect(addr1).burn(tokenId),
     )
       .to.emit(erc721, "Transfer")
       .withArgs(addr1.address, nullAddress, tokenId);
     await expect(erc721.connect(addr2).broadcastSelfTransfer(tokenId))
+      .to.be.revertedWithCustomError(erc721, "ERC721UniversalAlreadyTransferred")
+      .withArgs(tokenId);
+  });
+
+  it("broadcastSelfTransferBatch reverts on at least one burned asset", async function () {
+    const tokenId = buildTokenId("111", addr1.address);
+    const tokenId2 = buildTokenId("222", addr1.address);
+    await expect(
+      erc721.connect(addr1).burn(tokenId),
+    )
+      .to.emit(erc721, "Transfer")
+      .withArgs(addr1.address, nullAddress, tokenId);
+    await expect(erc721.connect(addr2).broadcastSelfTransferBatch([tokenId, tokenId2]))
       .to.be.revertedWithCustomError(erc721, "ERC721UniversalAlreadyTransferred")
       .withArgs(tokenId);
   });